### 2016-07-01 #35 @whedon
[REVIEW]: LAS: an integrated language analysis tool for multiple languages
**Submitting author:** @jiemakel (<a href="http://orcid.org/0000-0002-8366-8414">Eetu Mäkelä</a>)
**Repository:** <a href="https://github.com/jiemakel/las" target ="_blank">https://github.com/jiemakel/las</a>
**Version:** v1.4.8
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.160256" target="_blank">https://dx.doi.org/10.5281/zenodo.160256</a>
