### 2016-07-01 f67327d
add DOI to one paper that could be found
