+++ paper.bib
@@ -63,12 +63,23 @@
   doi = {10.1045/july2016-paakkonen}
 }
 
-@inproceedings{makela-sarpa-2014,
-  author =   {Eetu Mäkelä},
-  title =    {Combining a REST Lexical Analysis Web Service with SPARQL for Mashup Semantic Annotation from Text},
-  year = {2014},
-  booktitle = {Proceedings of the ESWC 2014 demonstration track, Springer-Verlag},
-  month = {May}
+@Inbook{makela-sarpa-2014,
+author="M{\"a}kel{\"a}, Eetu",
+editor="Presutti, Valentina
+and Blomqvist, Eva
+and Troncy, Raphael
+and Sack, Harald
+and Papadakis, Ioannis
+and Tordai, Anna",
+title="Combining a REST Lexical Analysis Web Service with SPARQL for Mashup Semantic Annotation from Text",
+bookTitle="The Semantic Web: ESWC 2014 Satellite Events: ESWC 2014 Satellite Events, Anissaras, Crete, Greece, May 25-29, 2014, Revised Selected Papers",
+year="2014",
+publisher="Springer International Publishing",
+address="Cham",
+pages="424--428",
+isbn="978-3-319-11955-7",
+doi="10.1007/978-3-319-11955-7_60",
+url="http://dx.doi.org/10.1007/978-3-319-11955-7_60"
 }
 
 @inproceedings{moshagenopen,
