### 2016-07-18 #35 @desilinguist
#### Installation

There's no specific installation section just a section called "Running". It would be nice if we could just rename that section to "Installation" even though there's nothing really to install since that's what people generally look for. And it would be nice to have that section contain the actual steps needed to run LAS on the example referenced later.

Another point is that the `las` binary that you download from the releases page needs to be chmod'ed 755 before it can run at least on my mac.

#### Functionality

I downloaded the file and tested on some Hindi text and it seems to work fine. However, as an aside, the web service inaccurately detects the same text as German for some reason (even though the underlying language detector says `hi` but for some reason it gets over-ruled by the language recognizer?

Another problem I had was that when I called `las recognize test.txt` on my Hindi source file instead of `las identify test.txt`, I get the following exception.

```java
10:15:03.846 [main] DEBUG com.optimaize.langdetect.LanguageDetectorImpl - ==> [DetectedLanguage[hi:0.8542185083632905], DetectedLanguage[mr:0.14570001916332004]]
Exception in thread "main" java.lang.IllegalArgumentException: in must not be null!
	at opennlp.tools.util.model.BaseModel.<init>(BaseModel.java:179)
	at opennlp.tools.tokenize.TokenizerModel.<init>(TokenizerModel.java:125)
	at fi.seco.lexical.combined.CombinedLexicalAnalysisService.getTokenizer(CombinedLexicalAnalysisService.java:125)
	at fi.seco.lexical.combined.CombinedLexicalAnalysisService.recognize(CombinedLexicalAnalysisService.java:617)
	at LASCommandLineTool$.recognize(LASCommandLineTool.scala:299)
	at LASCommandLineTool$$anonfun$main$9$$anonfun$apply$6.apply(LASCommandLineTool.scala:219)
	at LASCommandLineTool$$anonfun$main$9$$anonfun$apply$6.apply(LASCommandLineTool.scala:218)
	at scala.collection.immutable.List.foreach(List.scala:381)
	at LASCommandLineTool$$anonfun$main$9.apply(LASCommandLineTool.scala:218)
	at LASCommandLineTool$$anonfun$main$9.apply(LASCommandLineTool.scala:209)
	at scala.collection.immutable.List.foreach(List.scala:381)
	at LASCommandLineTool$.main(LASCommandLineTool.scala:209)
	at LASCommandLineTool.main(LASCommandLineTool.scala)
```
Is it simply because the LAS recognizer doesn't actually support Hindi (according to the help message)? If so, a more interpretable exception might be more useful here.

#### Performance

Although there are no specific claims made about the performance, the README does seem to imply that re-running after the first time should be much faster. I didn't really experience that. For example, running `las identify` on the same single sentence text file took about 25-30 seconds no matter how many times I re-ran it.

#### Documentation

This section seems lacking in general.

There isn't really a statement of need. It would be nice to have something that explains what the tool is useful for and also all of the modes of analysis it provides (recognition, identification etc.).

The package is self-contained so there's no need for dependencies, I guess. However, there's no information for people who are interested in contributing. There also don't seem to be any automated tests to run.

There's an example which is useful but it seems quite contrived. It would be nice to have actual examples for all of the analysis modes.

#### Software paper

The paper actually does have actual text motivating the need for this software. This text can be used to add the statement of need to the repository README.

#### Recommendation

Accept with minor changes. I think a more detailed README that explains the motivation, has better examples, and better installation/run instructions is all this package needs. It should also emphasize more prominently the fact that it's a wrapper around other existing projects.
