### 2016-07-19 #35 @pjotrp
This software is essentially a 300+ line script that wraps existing Java tools. I would like to ask the author to confirm **here** that he thinks this software is a valuable contribution to research. I want to point out that once published this work will be visible for a long time. Adding an example of having used this tool in research would be a good idea.
