### 2016-09-20 #35 @arfon
@desilinguist - This looks like this is in reasonable shape to accept. Can you confirm?
