### 2016-10-12 #35 @arfon
> I'm happy to accept this. Sorry for the delay.

✨ thanks for the update @desilinguist.
